Allow for using the same or different Route53 zones for both stacks
If using a single Route53 zone, assume that we want to prepend the subdomain.
Otherwise, assume that we want to use the same name as the zone.
In prod, you'll want to use separate zones (in separate accounts), but using
a single zone and account can be useful in pre-prod environments.