Use FQDN for Route53 A records
Use the FQDN for CTS and ZeroKMS APIs in the Route53 A records that
point to API GW. Terminating with "." tells CDK to use a fully qualified
domain name instead of making the name relative to the zone name.

Using the FQDN will work for both multi-account (prod) and single-account
(pre-prod) deployments.

See:
https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_route53.ARecord.html#recordname